move tests into existing `ReportTest`
